+++ src/controllers/tagControllers.js
@@ -28,7 +28,8 @@ export async function getTagByIdHandler(req, res, next) {
 
 export async function createTagHandler(req, res, next) {
 	try {
-		const tag = await createTag(req.body);
+		const tagData = { ...req.body, ownerId: req.user.id };
+		const tag = await createTag(tagData);
 		res.status(201).json(tag);
 	} catch (err) {
 		next(err);
@@ -38,6 +39,12 @@ export async function createTagHandler(req, res, next) {
 export async function updateTagHandler(req, res, next) {
 	try {
 		const id = Number(req.params.id);
+		const existingTag = await getTagById(id);
+		if (existingTag.ownerId !== req.user.id) {
+			const error = new Error('Forbidden: You are not the owner of this tag');
+			error.status = 403;
+			throw error;
+		}
 		const tag = await updateTag(id, req.body);
 		res.json(tag);
 	} catch (err) {
@@ -48,6 +55,12 @@ export async function updateTagHandler(req, res, next) {
 export async function deleteTagHandler(req, res, next) {
 	try {
 		const id = Number(req.params.id);
+		const existingTag = await getTagById(id);
+		if (existingTag.ownerId !== req.user.id) {
+			const error = new Error('Forbidden: You are not the owner of this tag');
+			error.status = 403;
+			throw error;
+		}
 		await deleteTag(id);
 		res.status(204).send();
 	} catch (err) {
